break

+++ test.py
@@ -4,7 +4,7 @@ result_of_random = simple_script.add_vars()
 print(result_of_random)
 
 def test_is_int():
-    is_int = isinstance(result_of_random, float)
+    is_int = isinstance(result_of_random, int)
     assert is_int is True
 
 def test_lower_than_13():
